+++ protoc_plugin/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 14.0.0
+
+* Breaking change: generated message classes now have a new instance method `createEmptyInstance`
+  that is used to support the new `toBuilder()` semantics of protobuf 0.11.0.
+  The generated code requires at least protobuf 0.11.0.
+
 ## 13.0.1
 
 * Add test for recursive merging and update protobuf dependency to 0.10.7.